requirements/dev.txt
-r prod.txt
black==24.1.1
django_debug_toolbar==4.3.0
